Check for invalid numbers
It's possible to enter invalid numbers into a number input (e.g. with
'-', '.', and 'e' characters). The event handler receives an empty
string in these cases, but we just check that the parsed int is valid.

This prevents a typo from clearing the user's input entirely (based on
the app's handling of null & NaN values).